requirements.txt
flask
transformers
torch
openai
python-dotenv
langchain
langchain_community